+++ daiv/codebase/models.py
@@ -9,8 +9,6 @@ from django_extensions.db.models import TimeStampedModel
 from langchain_core.documents import Document
 from pgvector.django import HnswIndex, VectorField
 
-from .conf import settings
-
 if TYPE_CHECKING:
     from .base import Repository
 
@@ -91,7 +89,8 @@ class CodebaseDocument(TimeStampedModel):
     source = models.CharField(max_length=256)
     description = models.TextField(blank=True)
     page_content = models.TextField()
-    page_content_vector = VectorField(dimensions=settings.EMBEDDINGS_DIMENSIONS)
+    # This will accept vectors of any dimension on first insert
+    page_content_vector = VectorField(dimensions=None)
     is_default_branch = models.BooleanField(default=False)
     metadata = models.JSONField(default=dict)