+++ CHANGELOG.md
@@ -9,11 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added support for `HuggingFace` embeddings models.
+- Added support for `HuggingFace` and `VoyageAI` embeddings models.
 
 ### Changed
 
 - Improved `README.md` to include information about how to setup a test project on local GitLab.
+- Command `update_index` now has a `--semantic-augmented-context` flag to conditionally generate semantic augmented context.
 
 ### Fixed